requirements.txt
absl-py
appdirs
apturl
argcomplete
asgiref
astunparse
attrs
autopep8
awscli
backcall
bcrypt
beautifulsoup4
beniget
blinker
boto3
botocore
Brlapi
Brotli
build
cachetools
certifi
cffi
cfn-flip
chardet
click
colorama
command-not-found
cryptography
cupshelpers
cycler
dbus-python
decorator
deepface
defer
distlib
distro
distro-info
dj-database-url
Django
django-cors-headers
django-cors-middleware
django-heroku
djangorestframework
dlib
docopt
docutils
duplicity
durationpy
evdev
face-recognition
face-recognition-models
fasteners
filelock
filetype
fire
Flask
flatbuffers
fonttools
fs
future
gast
gdown
google-auth
google-auth-oauthlib
google-pasta
grpcio
gunicorn
gyp
h5py
hjson
html5lib
httplib2
idna
importlib-metadata
ipython
itsdangerous
jedi
jeepney
Jinja2
jmespath
kappa
keras
keyring
kiwisolver
language-selector
launchpadlib
lazr.restfulclient
lazr.uri
libclang
lockfile
louis
lxml
lz4
macaroonbakery
Mako
Markdown
MarkupSafe
matplotlib
matplotlib-inline
monotonic
more-itertools
mpmath
mtcnn
mysql-connector-python
mysqlclient
netifaces
numpy
oauthlib
olefile
opencv-python
opt-einsum
packaging
pandas
paramiko
parso
pep517
pexpect
pickleshare
Pillow
pip-tools
pipenv
pipreqs
placebo
platformdirs
ply
prompt-toolkit
protobuf
psycopg2
psycopg2-binary
ptyprocess
pyasn1
pyasn1-modules
pycairo
pycodestyle
pycparser
pycups
Pygments
PyGObject
PyJWT
pymacaroons
PyNaCl
pynput
pyparsing
pyRFC3339
PySocks
python-apt
python-dateutil
python-debian
python-decouple
python-slugify
python-xlib
pythran
pytz
pyxdg
PyYAML
reportlab
requests
requests-oauthlib
retina-face
roman
rsa
s3transfer
scapy
scipy
SecretStorage
six
soupsieve
sqlparse
sympy
systemd-python
tabulate
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
tensorflow-io-gcs-filesystem
termcolor
text-unidecode
toml
tomli
tqdm
traitlets
troposphere
typing-extensions
tzdata
ubuntu-advantage-tools
ubuntu-drivers-common
ufoLib2
ufw
unattended-upgrades
unicodedata2
urllib3
usb-creator
uuid
virtualenv
virtualenv-clone
wadllib
wcwidth
webencodings
Werkzeug
whitenoise
wrapt
wsgi-request-logger
xdg
xkit
yarg
zappa
zipp